feat: migrate to santuario c14n
instead of using a patched fork of apache c14n from 2013

What to do with this branch? I think I want to remove the block
parameter from the JRuby API, because

a) it didn't previously work the same as CRuby/libxml2 (see commit on
this branch that introduces block test coverage that fails)

b) with the mainline c14n library we can't hack it to support even the
broken callback semantics and would be forced to implement something
complicated ourselves

Given I'm not confident that C14n is widely-used anyway, removing a
misfeature like this doesn't seem like a controversial decision.

So, questions to the JRuby contributors:

Q1: Is this a reasonable decision?
Q2a: Are we OK with all these new dependencies getting pulled in?
Q2b: There's some Log4J warning messages being emitted at runtime that
I'm not sure whether or how to quash. Any ideas?